Removes Runner in favor of a RequestInterceptor class
