+++ lib/request_interceptor/version.rb
@@ -1,3 +1,3 @@
-module RequestInterceptor
+class RequestInterceptor
   VERSION = "0.1.0"
 end
